extract convertToReadable function and place logic in a util file called helperFunctions.js to be utilized across other files such as the Sponsors Component
