+++ frontend/src/util/helperFunctions.js
@@ -1,0 +1,13 @@
+//function takes array value and seperates elements by commas and spaces
+export const convertToReadable = array => {
+  if (array) {
+    let newArray = array;
+
+    newArray = newArray.map(item => {
+      return item = " " + item;
+    });
+
+    let convertedArray = newArray.toString();
+    return convertedArray
+  }
+}
